Prevent legacy schema manager from dropping unrecognized tables.

+++ library/PEAR/MDB2/Schema.php
@@ -83,12 +83,13 @@ class MDB2_Schema extends PEAR
 
     var $options = array(
         'fail_on_invalid_names' => true,
-        'dtd_file' => false,
-        'valid_types' => array(),
-        'force_defaults' => true,
-        'parser' => 'MDB2_Schema_Parser',
-        'writer' => 'MDB2_Schema_Writer',
-        'validate' => 'MDB2_Schema_Validate'
+        'dtd_file'              => false,
+        'valid_types'           => array(),
+        'force_defaults'        => true,
+        'parser'                => 'MDB2_Schema_Parser',
+        'writer'                => 'MDB2_Schema_Writer',
+        'validate'              => 'MDB2_Schema_Validate',
+        'drop_missing_tables'   => false
     );
 
     // }}}
@@ -1404,7 +1405,9 @@ class MDB2_Schema extends PEAR
                     $changes['tables'] = MDB2_Schema::arrayMergeClobber($changes['tables'], $change);
                 }
             }
-            if (!empty($previous_definition['tables']) && is_array($previous_definition['tables'])) {
+
+            if (!empty($previous_definition['tables'])
+                && is_array($previous_definition['tables'])) {
                 foreach ($previous_definition['tables'] as $table_name => $table) {
                     if (empty($defined_tables[$table_name])) {
                         $changes['tables']['remove'][$table_name] = true;
@@ -1945,8 +1948,10 @@ class MDB2_Schema extends PEAR
                 $alterations++;
             }
         }
-
-        if (!empty($changes['remove']) && is_array($changes['remove'])) {
+ 
+        if ($this->options['drop_missing_tables']
+            && !empty($changes['remove'])
+            && is_array($changes['remove'])) {
             foreach ($changes['remove'] as $table_name => $table) {
                 $result = $this->db->manager->dropTable($table_name);
                 if (PEAR::isError($result)) {
@@ -2148,9 +2153,15 @@ class MDB2_Schema extends PEAR
                 }
             }
             if (!empty($changes['tables']['remove']) && is_array($changes['tables']['remove'])) {
-                foreach ($changes['tables']['remove'] as $table_name => $table) {
-                    $this->db->debug("$table_name:", __FUNCTION__);
-                    $this->db->debug("\tRemoved table '$table_name'", __FUNCTION__);
+                if ($this->options['drop_missing_tables']) {
+                    foreach ($changes['tables']['remove'] as $table_name => $table) {
+                        $this->db->debug("$table_name:", __FUNCTION__);
+                        $this->db->debug("\tRemoved table '$table_name'", __FUNCTION__);
+                    }
+                } else {
+                    foreach ($changes['tables']['remove'] as $table_name => $table) {
+                        $this->db->debug("\tDeprecated table '$table_name' left as is", __FUNCTION__);
+                    }
                 }
             }
             if (!empty($changes['tables']['change']) && is_array($changes['tables']['change'])) {
